+++ packages/core/src/jsx-namespace.ts
@@ -1,10 +1,10 @@
-import { SxProps } from './types'
+import { SxProp } from './types'
 
-type WithConditionalSxProps<P> = 'className' extends keyof P
-  ? P extends { className?: string }
-    ? Omit<P, keyof SxProps> & SxProps
-    : Omit<P, keyof SxProps>
-  : Omit<P, keyof SxProps>
+type WithConditionalSxProp<P> = 'className' extends keyof P
+  ? string extends P['className']
+    ? P & SxProp
+    : P
+  : P
 
 type ReactJSXElement = JSX.Element
 type ReactJSXElementClass = JSX.ElementClass
@@ -22,13 +22,20 @@ export declare namespace ThemeUIJSX {
     extends ReactJSXElementAttributesProperty {}
   export interface ElementChildrenAttribute
     extends ReactJSXElementChildrenAttribute {}
-  export type LibraryManagedAttributes<C, P> = WithConditionalSxProps<P> &
+  export type LibraryManagedAttributes<C, P> = WithConditionalSxProp<P> &
+    // We are not removing incompatible `sx` props, because touching this breaks
+    // inference in generic components.
+    // Yes, we steal any prop called `sx` at runtime, but we
+    // can't represent it on type level without breaking compatibility with
+    // our own Field, react-hook-form, and a bunch of other generic components.
+    // Don't touch ReactJSXLibraryManagedAttributes or you'll spend hours
+    // debugging and entirely spoil your day.
     ReactJSXLibraryManagedAttributes<C, P>
   export interface IntrinsicAttributes extends ReactJSXIntrinsicAttributes {}
   export interface IntrinsicClassAttributes<T>
     extends ReactJSXIntrinsicClassAttributes<T> {}
   export type IntrinsicElements = {
     [K in keyof ReactJSXIntrinsicElements]: ReactJSXIntrinsicElements[K] &
-      SxProps
+      SxProp
   }
 }