+++ packages/docs/src/components/code.js
@@ -1,5 +1,5 @@
 /** @jsx jsx */
-import { jsx, Styled, Text } from 'theme-ui'
+import { jsx, Themed, Text } from 'theme-ui'
 import Prism from '@theme-ui/prism'
 import { LiveProvider, LiveEditor, LivePreview, LiveError } from 'react-live'
 import * as themeUI from 'theme-ui'
@@ -42,12 +42,12 @@ const images = {
   flatiron:
     'https://images.unsplash.com/photo-1520222984843-df35ebc0f24d?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=1080&fit=max&ixid=eyJhcHBfaWQiOjF9',
   logo:
-    'https://raw.githubusercontent.com/system-ui/theme-ui/master/packages/docs/static/logo.png',
+    'https://raw.githubusercontent.com/system-ui/theme-ui/stable/packages/docs/static/logo.png',
 }
 
 const scope = {
   ...themeUI,
-  Link: props => {
+  Link: (props) => {
     if (props.activeClassName)
       return <span className={props.activeClassName} {...props} />
     return <span {...props} sx={{ cursor: 'pointer' }} />
@@ -56,7 +56,7 @@ const scope = {
   images,
 }
 
-const transformCode = src => `/** @jsx jsx */\n<>${src}</>`
+const transformCode = (src) => `/** @jsx jsx */\n<>${src}</>`
 
 const liveTheme = { styles: [] }
 
@@ -83,7 +83,7 @@ export const LiveCode = ({ children, preview, xray }) => {
         sx={{
           p: 3,
           variant: xray ? 'styles.xray' : null,
-          border: t => `1px solid ${t.colors.muted}`,
+          border: (t) => `1px solid ${t.colors.muted}`,
         }}>
         <LivePreview />
         <LiveError
@@ -97,18 +97,18 @@ export const LiveCode = ({ children, preview, xray }) => {
           }}
         />
       </div>
-      <Styled.pre
+      <Themed.pre
         sx={{
           mt: 0,
           mb: 3,
         }}>
         <LiveEditor padding={0} />
-      </Styled.pre>
+      </Themed.pre>
     </LiveProvider>
   )
 }
 
-const Code = props => {
+const Code = (props) => {
   if (props.live) {
     return <LiveCode {...props} />
   }
@@ -131,7 +131,7 @@ const Code = props => {
       </section>
     )
   }
-  return <div {...props} />
+  return <Prism {...props} />
 }
 
 export default Code