+++ packages/gatsby-theme-style-guide/src/template.js
@@ -4,7 +4,7 @@ import Header from './header'
 import Typography from './typography'
 import Colors from './colors'
 
-export default props => (
+export default (props) => (
   <Layout>
     <Header />
     <Typography />