+++ CHANGELOG.md
@@ -1,6 +1,29 @@
 # Changelog
 
-## v0.5.0 UNRELEASED
+## v0.6.0 UNRELEASED
+
+- Extract objects with nested variant props. Issue #1357
+- Add ability for MDX styling, and fix mdx table align styles. Issue #654
+- Remove recursive default values from CSS custom properties. PR #1327
+- Support a `"default"` key for object in scales. PR #951
+- Render extra Embed props onto `iframe` tag instead of wrapping `div`. Issue #966, PR #1122
+
+## v0.6.0-alpha.1
+
+- Switches from lodash.kebabCase to alternative package ([param-case](/blakeembrey/change-case/tree/master/packages/param-case)) per [official Lodash documentation](https://lodash.com/per-method-packages). PR #1304
+- Rebuilds Prism preset with latest upstream theme changes. PR #1304
+- Fix: Preserve order of variant expansion props. PR #1326
+  _(bug introduced in 0.5.0-alpha.1)_
+
+## v0.6.0-alpha.0
+
+- **BREAKING**: Rename `Styled` component to `Themed`. PR #1323
+- **BREAKING**: Make Text component use `span` instead of `div`
+- **breaking TypeScript**: Renamed and removed types. PR #1308
+  - `SxProps` to `SxProp`.
+  - `SxStyleProp`, an alias for `ThemeUIStyleObject` removed. Use `ThemeUIStyleObject` instead.
+- Fix: Add `sx` props types to all props accepting `className`. PR #1308
+- Fix WithPoorAsProp to work with ComponentProps utility type. PR #1308
 
 ## v0.5.0-alpha.2 2020-11-30
 
@@ -10,11 +33,12 @@
 
 - Bump React peerDependency to `"^16.14.0 || ^17.0.0"`.
 - Support automatic JSX runtime. Issue #1160, PR #1237
+- Bump React peerDependency to `"^16.14.0 || ^17.0.0"`.
 - Apply variant styles _before_ responsive styles. Issues #1030, and #720, PR #1273
 
 ## v0.5.0-alpha.0 2020-11-20
 
-- BREAKING: Upgrade to Emotion 11, and `csstype` 3. PR #1261
+- **BREAKING**: Upgrade to Emotion 11, and `csstype` 3. PR #1261
   - We are now depending on `@emotion/react@11` instead of `@emotion/core@10`
   - `sx` prop types are still global, and we opt in for Emotion `css` prop types (This will change in the future.)
   - Refer to [Emotion 11 release notes](https://emotion.sh/docs/emotion-11) for more information.